+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/svelte
 
+## 3.0.0-canary.8
+
+### Patch Changes
+
+- Updated dependencies [a847c3e]
+- Updated dependencies [5d142ab]
+- Updated dependencies [cb9c9e4]
+- Updated dependencies [8aa9e20]
+- Updated dependencies [b32e192]
+  - ai@5.0.0-canary.9
+  - @ai-sdk/provider-utils@3.0.0-canary.8
+
 ## 3.0.0-canary.7
 
 ### Patch Changes